add some debug prints for cog issue...
