+++ scripts/deforum_helpers/render.py
@@ -49,6 +49,7 @@ from .prompt import prepare_prompt
 from modules.shared import opts, cmd_opts, state, sd_model
 from modules import lowvram, devices, sd_hijack
 from .RAFT import RAFT
+from .general_utils import stable_hash
 
 from deforum_api import JobStatusTracker
 
@@ -542,14 +543,20 @@ def render_animation(args, anim_args, video_args, parseq_args, loop_args, contro
 
         # optical flow redo before generation
         if optical_flow_redo_generation != 'None' and prev_img is not None and strength > 0:
-            print(f"Optical flow redo is diffusing and warping using {optical_flow_redo_generation} optical flow before generation.")
             stored_seed = args.seed
-            args.seed = random.randint(0, 2 ** 32 - 1)
+            if anim_args.optical_flow_redo_generation == "Random":
+                args.seed = random.randint(0, 2 ** 32 - 1)
+            elif anim_args.optical_flow_redo_generation == "Reproducible":
+                args.seed = stable_hash(args.seed)
+
+            print(f"Optical flow redo is diffusing and warping using {optical_flow_redo_generation} and seed {args.seed} optical flow before generation.")                
+
             disposable_image = generate(args, keys, anim_args, loop_args, controlnet_args, root, parseq_adapter, frame_idx, sampler_name=scheduled_sampler_name)
             disposable_image = cv2.cvtColor(np.array(disposable_image), cv2.COLOR_RGB2BGR)
             disposable_flow = get_flow_from_images(prev_img, disposable_image, optical_flow_redo_generation, raft_model)
             disposable_image = cv2.cvtColor(disposable_image, cv2.COLOR_BGR2RGB)
             disposable_image = image_transform_optical_flow(disposable_image, disposable_flow, redo_flow_factor)
+            save_image(disposable_image, 'PIL', os.path.join(args.outdir, f"flow_generation_{root.timestring}_{frame_idx:09}.png"), args, video_args, root)
             args.seed = stored_seed
             root.init_sample = Image.fromarray(disposable_image)
             del (disposable_image, disposable_flow, stored_seed)
