+++ scripts/deforum_helpers/video_audio_utilities.py
@@ -474,10 +474,10 @@ def render_preview(args, anim_args, video_args, root, frame_idx, last_preview_fr
     is_preview_frame = (frame_idx % preview_interval_frames) == 0 or (frame_idx - last_preview_frame) >= preview_interval_frames
     is_close_to_end = frame_idx >= (anim_args.max_frames-1)
 
-    debug_print(f"render preview video: frame_idx={frame_idx} preview_interval_frames={preview_interval_frames} anim_args.max_frames={anim_args.max_frames} is_preview_on={is_preview_on} is_preview_frame={is_preview_frame} is_close_to_end={is_close_to_end} ")
+    print(f"render preview video: frame_idx={frame_idx} preview_interval_frames={preview_interval_frames} anim_args.max_frames={anim_args.max_frames} is_preview_on={is_preview_on} is_preview_frame={is_preview_frame} is_close_to_end={is_close_to_end} ")
 
     if not is_preview_on or not is_preview_frame or is_close_to_end:
-        debug_print(f"No preview video on frame {frame_idx}.")
+        print(f"No preview video on frame {frame_idx}.")
         return last_preview_frame
     
     f_location, f_crf, f_preset = get_ffmpeg_params() # get params for ffmpeg exec
